# Make free-connection stack a LIFO.