```
benchmark                              old ns/op     new ns/op     delta
BenchmarkPoolGetPut10Conns-4           209           385           +84.21%
BenchmarkPoolGetPut100Conns-4          215           389           +80.93%
BenchmarkPoolGetPut1000Conns-4         236           388           +64.41%
BenchmarkPoolGetReplace10Conns-4       319           376           +17.87%
BenchmarkPoolGetReplace100Conns-4      294           409           +39.12%
BenchmarkPoolGetReplace1000Conns-4     294           410           +39.46%

benchmark                              old allocs     new allocs     delta
BenchmarkPoolGetPut10Conns-4           0              0              +0.00%
BenchmarkPoolGetPut100Conns-4          0              0              +0.00%
BenchmarkPoolGetPut1000Conns-4         0              0              +0.00%
BenchmarkPoolGetReplace10Conns-4       3              3              +0.00%
BenchmarkPoolGetReplace100Conns-4      3              3              +0.00%
BenchmarkPoolGetReplace1000Conns-4     3              3              +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkPoolGetPut10Conns-4           0             0             +0.00%
BenchmarkPoolGetPut100Conns-4          0             0             +0.00%
BenchmarkPoolGetPut1000Conns-4         0             0             +0.00%
BenchmarkPoolGetReplace10Conns-4       32            32            +0.00%
BenchmarkPoolGetReplace100Conns-4      32            32            +0.00%
```